renaming xc2scope.py to x2cscope.py
