+++ pyx2cscope/examples/SFR_Example.py
@@ -5,7 +5,7 @@ import time
 
 from pyx2cscope.utils import get_com_port, get_elf_file_path
 
-from pyx2cscope.xc2scope import X2CScope
+from pyx2cscope.x2cscope import X2CScope
 
 # Configuration for serial port communication
 serial_port = get_com_port()  # Select COM port
