+++ pyx2cscope/examples/testingArray.py
@@ -3,7 +3,7 @@ import logging
 
 import matplotlib.pyplot as plt
 from pyx2cscope.utils import get_com_port, get_elf_file_path
-from pyx2cscope.xc2scope import X2CScope
+from pyx2cscope.x2cscope import X2CScope
 
 # Set up logging
 logging.basicConfig(
