+++ pyx2cscope/gui/web/static/style.css
@@ -12,3 +12,7 @@ a { color: inherit; }
 html {
     font-size: 12px;
 }
+
+.chartView {
+height: 50vh;
+}
