+++ pyx2cscope/gui/web/templates/scope_view.html
@@ -1,5 +1,4 @@
-<canvas id="scopeChart" width="400" height="200"></canvas>
-<br>
+        <canvas class="chartView" id="scopeChart" width="400" height="200"></canvas>
 <div class="row">
     <div class="col-md-4">
         <div class="card">
