+++ tests/test_pyx2cscope_class.py
@@ -3,7 +3,7 @@ import os
 
 import pytest
 
-from pyx2cscope.xc2scope import X2CScope
+from pyx2cscope.x2cscope import X2CScope
 from tests import data
 from tests.utils.serial_stub import fake_serial
